testDefaultPort
